Address account identity review feedback
